+++ src/formatterExecutor.ts
@@ -4,10 +4,22 @@ export default function (document: TextDocument, options: FormattingOptions): Te
 {
     var completions: TextEdit[] = [];
 
+    // console.time("Format");
+    // console.time("Indentation");
     completions.push(...require('./formatters/indentation').format(document, options));
-    completions.push(...require('./formatters/ifOperatorSpacing').format(document, options));
+    // console.timeEnd("Indentation");
+    // console.time("Argument commas");
     completions.push(...require('./formatters/argumentCommas').format(document, options));
+    // console.timeEnd("Argument commas");
+    // console.time("Semicolons");
     completions.push(...require('./formatters/semicolons').format(document, options));
-
-    return completions
+    // console.timeEnd("Semicolons");
+    // console.time("Braces");
+    completions.push(...require('./formatters/bracesPlacer').format(document, options));
+    // console.timeEnd("Braces");
+    // console.time("Input Output Aligner");
+    completions.push(...require('./formatters/inputOutputAligner').format(document, options));
+    // console.timeEnd("Input Output Aligner");
+    // console.timeEnd("Format");
+    return completions;
 }