+++ CHANGELOG.md
@@ -1,6 +1,25 @@
+# 1.1.0
+
+## Added
+
+- Brace placement
+- Input Output alignment
+- Status icon in the bottom bar (Click to go to extension setttings)
+- Settings
+- Suggestion for Simpl+ by mwgustin
+
+## Changed
+
+- Semicolons will now not be placed in block comments
+
+## Removed
+
+- Statement operator spacing (I couldn't make it work)
+
 # 1.0.0
 
 ## Added
+
 - Formatter to add semicolons.
 - Formatter to add a space to commas in arguments.
 - Formatter to add a space to operators in if statements.